AI-Resume_Screening-app-main/requirements.txt
Flask
pdfplumber
langchain
langchain_groq
werkzeug
python-dotenv
